Fixed #58: New angle step user option
Angle step was a fixed constant in CalcRotationAngle() set to 0.1 degrees, now it's a user option (default value as before 0.1). Can be used e.g. with a bigger step to quickly detect angle when accuracy does not need to be the best, or go lower than default for better accuracy but with much longer running time. Or for automated 2-step run: first big step and big max angle to find rough skew, and then